+++ docs/reference/changelog.rst
@@ -15,7 +15,24 @@ Changelog
 
     For a complete list of commits, check out the `1.2.3`_ release on GitHub.
 
-7.3.0 (unreleased)
+7.4.0 (2026-09-25)
+------------------
+
+Application
+===========
+
+- A new ``AppMetadata.enable_build_slices`` setting enables support for build slices.
+
+Project
+=======
+
+- A new top-level ``build-slices`` key can be used to define build slices that are
+  used by many parts, to avoid repetition.
+
+For a complete list of commits, check out the `7.4.0`_ release on GitHub.
+
+
+7.3.0 (2026-09-23)
 ------------------
 
 Application
@@ -24,11 +41,18 @@ Application
 - A new ``AppMetadata.allow_spread_yaml`` setting controls whether the ``test`` command
   can fall back to the deprecated ``spread.yaml``.
 
+- Artifact digests are now computed by reading files in chunks instead of
+  loading them entirely into memory.
+
 Commands
 ========
 
 - The ``spread.yaml`` file for the ``test`` command has been deprecated in favor
   of ``<app-name>-test.yaml``.
+
+- The ``init`` command now supports Git features.
+
+- The ``pack`` command no longer packages projects that define no parts.
 
 Services
 ========
@@ -36,13 +60,14 @@ Services
 - The testing service now handles multiple artifacts by defining variables
   ``CRAFT_ARTIFACT_<name>`` in the Spread test environment.
 
+- The testing service generates the Launchpad test back-end.
+
 - Add the Launchpad testing farm back-end to the testing service.
 
-Bug fixes
-=========
+- The testing service will proxy runner traffic through a fetch-service session
+  when ``CRAFT_USE_EXTERNAL_FETCH_SERVICE`` is set.
 
-- Hosts reporting ``armv8l`` (an ARMv8 processor running a 32-bit user space) are
-  now correctly identified as ``armhf``.
+For a complete list of commits, check out the `7.3.0`_ release on GitHub.
 
 
 7.2.1 (2026-09-01)
@@ -1460,3 +1485,5 @@ For a complete list of commits, check out the `2.7.0`_ release on GitHub.
 .. _7.1.0: https://github.com/canonical/craft-application/releases/tag/7.1.0
 .. _7.2.0: https://github.com/canonical/craft-application/releases/tag/7.2.0
 .. _7.2.1: https://github.com/canonical/craft-application/releases/tag/7.2.1
+.. _7.3.0: https://github.com/canonical/craft-application/releases/tag/7.3.0
+.. _7.4.0: https://github.com/canonical/craft-application/releases/tag/7.4.0